Moved /users/me to core.routes; reusing helpers
