+++ donut/modules/directory_search/templates/view_user.html
@@ -9,7 +9,7 @@
 				{{ user['middle_name'] or '' }}
 				{{ user['last_name'] }}
 				{% if is_me %}
-					<a href='{{ url_for("directory_search.edit_user") }}' style='float: right; padding-right: 20px'>Update</a>
+					<a href='{{ url_for("core.edit_user") }}' style='float: right; padding-right: 20px'>Update</a>
 				{% endif %}
 			</h2>
 			{% if user['image'] %}
