+++ donut/templates/includes/header.html
@@ -68,7 +68,7 @@
                     </li>
                     {% else %}
                     <li>
-                        <a href="{{ url_for('directory_search.my_directory_page') }}">
+                        <a href="{{ url_for('core.my_directory_page') }}">
                             Greetings, {{ session['username'] }}
                         </a>
                     </li>
