+++ tests/modules/directory_search/test_directory_search.py
@@ -6,6 +6,7 @@ import flask
 import pytest
 from donut.testing.fixtures import client
 from donut import app
+import donut.modules.core.helpers as core_helpers
 from donut.modules.directory_search import helpers
 from donut.modules.directory_search import routes
 
@@ -19,8 +20,6 @@ def test_hidden_fields(client):
 
 
 def test_get_user(client):
-    assert helpers.get_user_id('csander') == 3
-    assert helpers.get_user_id('whodis') == 0
     user_data = helpers.get_user(helpers.get_user_id('csander'))
     assert user_data == {
         'address':
@@ -89,7 +88,7 @@ def test_get_user(client):
     user_data2 = helpers.get_user(helpers.get_user_id('dqu'))
     assert user_data2['image'] == 0
     assert 'gender_string' not in user_data2
-    assert 'hometown_string' not in user_data2
+    assert not user_data2['hometown_string']
     assert 'phone_string' not in user_data2
     user_data3 = helpers.get_user(helpers.get_user_id('reng'))
     assert user_data3['phone_string'] == '+11234567890'
@@ -115,7 +114,7 @@ def test_name_query(client):
 def test_image(client):
     user_id = helpers.get_user_id('csander')
     assert helpers.get_image(user_id) == ('png', b'NOT_A_REAL_IMAGE')
-    helpers.set_image(user_id, 'jpg', 'FAKE_JPG')
+    core_helpers.set_image(user_id, 'jpg', 'FAKE_JPG')
     assert helpers.get_image(user_id) == ('jpg', b'FAKE_JPG')
     with pytest.raises(Exception):
         helpers.get_image(helpers.get_user_id('dqu'))
@@ -322,18 +321,18 @@ def test_value_lists(client):
 
 def test_preferred_name(client):
     user_id = helpers.get_user_id('csander')
-    assert helpers.get_preferred_name(user_id) == 'Cleb'
-    helpers.set_preferred_name(user_id, 'Belac')
-    assert helpers.get_preferred_name(user_id) == 'Belac'
-    assert helpers.get_preferred_name(helpers.get_user_id('reng')) == ''
+    assert core_helpers.get_preferred_name(user_id) == 'Cleb'
+    core_helpers.set_member_field(user_id, 'preferred_name', 'Belac')
+    assert core_helpers.get_preferred_name(user_id) == 'Belac'
+    assert core_helpers.get_preferred_name(helpers.get_user_id('reng')) == ''
 
 
 def test_gender(client):
     user_id = helpers.get_user_id('csander')
-    assert helpers.get_gender(user_id) == 'Male'
-    helpers.set_gender(user_id, 'new_gender')
-    assert helpers.get_gender(user_id) == 'new_gender'
-    assert helpers.get_gender(helpers.get_user_id('dqu')) == ''
+    assert core_helpers.get_gender(user_id) == 'Male'
+    core_helpers.set_member_field(user_id, 'gender_custom', 'new_gender')
+    assert core_helpers.get_gender(user_id) == 'new_gender'
+    assert core_helpers.get_gender(helpers.get_user_id('dqu')) == ''
 
 
 #Routes
@@ -345,7 +344,7 @@ def test_search_page(client):
 def test_my_page(client):
     with client.session_transaction() as sess:
         sess['username'] = 'csander'
-    res = client.get(flask.url_for('directory_search.my_directory_page'))
+    res = client.get(flask.url_for('core.my_directory_page'))
     assert res.status_code == 302
     assert res.headers['location'] == flask.url_for(
         'directory_search.view_user', user_id=3)
@@ -362,8 +361,7 @@ def test_view_user(client):
 def test_edit_page(client):
     with client.session_transaction() as sess:
         sess['username'] = 'csander'
-    assert client.get(
-        flask.url_for('directory_search.edit_user')).status_code == 200
+    assert client.get(flask.url_for('core.edit_user')).status_code == 200
 
 
 def test_set_name(client):
@@ -371,7 +369,7 @@ def test_set_name(client):
         sess['username'] = 'csander'
     assert helpers.get_user(3)['preferred_name'] == 'Belac'
     res = client.post(
-        flask.url_for('directory_search.set_name'), data={'name': 'Clb'})
+        flask.url_for('core.set_name'), data={'name': 'Clb'})
     assert res.status_code == 302
     assert res.headers['location'] == flask.url_for(
         'directory_search.view_user', user_id=3)
@@ -383,7 +381,7 @@ def test_set_name(client):
         sess['username'] = 'csander'
     assert helpers.get_user(3)['gender_string'] == 'new_gender'
     res = client.post(
-        flask.url_for('directory_search.set_gender'), data={'gender': 'Male'})
+        flask.url_for('core.set_gender'), data={'gender': 'Male'})
     assert res.status_code == 302
     assert res.headers['location'] == flask.url_for(
         'directory_search.view_user', user_id=3)
